Restructure tests to be outside of src and update sonar pointer
