+++ package.json
@@ -7,9 +7,9 @@
     "start": "node dist/bin/www.js",
     "start:watch": "NODE_ENV=development nodemon",
     "build": "NODE_ENV=production tsc && cp -r src/views dist/ && gulp static",
-    "lint": "eslint 'src/*.ts' 'src/**/*.ts' ./src/'test/*.ts' './src/test/**/*.ts'",
-    "lint:fix": "eslint 'src/*.ts' 'src/**/*.ts' './src/test/*.ts' './src/test/**/*.ts' --fix",
-    "test": "nyc mocha **/test/**/*.unit.ts",
+    "lint": "eslint 'src/*.ts' 'src/**/*.ts' ./'test/*.ts' './test/**/*.ts'",
+    "lint:fix": "eslint 'src/*.ts' 'src/**/*.ts' './test/*.ts' './test/**/*.ts' --fix",
+    "test": "nyc mocha test/**/*.unit.ts",
     "test:coverage": "nyc npm test",
     "analyse-code": "sonar-scanner"
   },
