+++ sonar-project.properties
@@ -6,6 +6,6 @@ sonar.password=${env.CODE_ANALYSIS_PASSWORD}
 
 sonar.projectName=search-web
 sonar.sources=src
-sonar.tests=src
+sonar.tests=test
 sonar.test.inclusions=**/test/**/*.unit.ts
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
