+++ test/client/apiclient.spec.unit.ts
@@ -4,12 +4,12 @@ import Resource from "@companieshouse/api-sdk-node/dist/services/resource";
 import { CompaniesResource } from "@companieshouse/api-sdk-node/dist/services/search/alphabetical-search/types";
 import { CompaniesResource as DissolvedCompanyResource } from "@companieshouse/api-sdk-node/dist/services/search/dissolved-search/types";
 import { CompaniesResource as AdvancedCompanyResource } from "@companieshouse/api-sdk-node/dist/services/search/advanced-search/types";
-import { getCompanies, getDissolvedCompanies, getAdvancedCompanies } from "../../client/apiclient";
+import { getCompanies, getDissolvedCompanies, getAdvancedCompanies } from "../../src/client/apiclient";
 import AlphabeticalSearchService from "@companieshouse/api-sdk-node/dist/services/search/alphabetical-search/service";
 import DissolvedSearchService from "@companieshouse/api-sdk-node/dist/services/search/dissolved-search/service";
 import AdvancedSearchService from "@companieshouse/api-sdk-node/dist/services/search/advanced-search/service";
 import { createDummyAdvancedSearchParams } from "../../test/MockUtils/advanced-search/mock.util";
-import { AdvancedSearchParams } from "../../model/advanced.search.params";
+import { AdvancedSearchParams } from "../../src/model/advanced.search.params";
 
 const mockResponse: Resource<CompaniesResource> = {
     httpStatusCode: 200,
