+++ test/controllers/alphabetical-search/search.controller.spec.unit.ts
@@ -3,7 +3,7 @@ import sinon from "sinon";
 import chai from "chai";
 import ioredis from "ioredis";
 import cheerio from "cheerio";
-import * as apiClient from "../../../client/apiclient";
+import * as apiClient from "../../../src/client/apiclient";
 import { signedInSession } from "../../MockUtils/redis.mocks";
 
 const sandbox = sinon.createSandbox();
@@ -14,7 +14,7 @@ describe("search.controller.spec.unit", () => {
     beforeEach((done) => {
         sandbox.stub(ioredis.prototype, "connect").returns(Promise.resolve());
         sandbox.stub(ioredis.prototype, "get").returns(Promise.resolve(signedInSession));
-        testApp = require("../../../../src/app").default;
+        testApp = require("../../../src/app").default;
         done();
     });
