+++ test/routes/routes.spec.unit.ts
@@ -11,7 +11,7 @@ describe("routes.spec.unit", () => {
         sandbox.stub(ioredis.prototype, "connect").returns(Promise.resolve());
         sandbox.stub(ioredis.prototype, "get").returns(Promise.resolve(signedInSession));
 
-        testApp = require("../../../src/app").default;
+        testApp = require("../../src/app").default;
         done();
     });
